Re-arrange some files
