+++ packages/wiggle/src/WiggleTrack/model.js
@@ -10,7 +10,7 @@ import {
 import { blockBasedTrackModel } from '@gmod/jbrowse-plugin-linear-genome-view'
 import { autorun } from 'mobx'
 import { addDisposer, getRoot, getSnapshot, types } from 'mobx-state-tree'
-import { getNiceDomain } from '../DensityRenderer/util'
+import { getNiceDomain } from '../util'
 import WiggleTrackComponent from './components/WiggleTrackComponent'
 
 // using a map because it preserves order
